trigger and VSS hw integrated testing #1668

+++ java_console/io/src/main/java/com/rusefi/binaryprotocol/BinaryProtocol.java
@@ -171,7 +171,7 @@ public class BinaryProtocol implements BinaryProtocolCommands {
         };
     }
 
-    public static void sleep(int millis) {
+    public static void sleep(long millis) {
         try {
             Thread.sleep(millis);
         } catch (InterruptedException e) {
